reorg v1/sinks/kafka/ modules
