+++ rust/capture/src/v1/sinks/constants.rs
@@ -1,0 +1,3 @@
+use std::time::Duration;
+
+pub(crate) const DEFAULT_PRODUCE_TIMEOUT: Duration = Duration::from_secs(30);
