+++ rust/capture/src/v1/sinks/types.rs
@@ -2,10 +2,7 @@ use std::borrow::Cow;
 use std::collections::HashMap;
 use std::fmt;
 
-use chrono::{DateTime, Utc};
-use rdkafka::error::RDKafkaErrorCode;
-
-use crate::v1::sinks::kafka::producer::ProduceError;
+use crate::v1::sinks::kafka::types::KafkaResult;
 
 /// Kafka topic routing for a processed event.
 /// `Drop` means the event should not be produced at all.
@@ -18,28 +15,6 @@ pub enum Destination {
     Dlq,
     Custom(String),
     Drop,
-}
-
-// ---------------------------------------------------------------------------
-// error_code_tag
-// ---------------------------------------------------------------------------
-
-/// Stable, low-cardinality snake_case tag for an RDKafkaErrorCode.
-/// Usable anywhere -- producer, sink, handler, logging.
-pub fn error_code_tag(code: RDKafkaErrorCode) -> &'static str {
-    match code {
-        RDKafkaErrorCode::QueueFull => "queue_full",
-        RDKafkaErrorCode::MessageSizeTooLarge => "message_size_too_large",
-        RDKafkaErrorCode::MessageTimedOut => "message_timed_out",
-        RDKafkaErrorCode::UnknownTopicOrPartition => "unknown_topic_or_partition",
-        RDKafkaErrorCode::TopicAuthorizationFailed => "topic_authorization_failed",
-        RDKafkaErrorCode::ClusterAuthorizationFailed => "cluster_authorization_failed",
-        RDKafkaErrorCode::InvalidMessage => "invalid_message",
-        RDKafkaErrorCode::InvalidMessageSize => "invalid_message_size",
-        RDKafkaErrorCode::NotLeaderForPartition => "not_leader_for_partition",
-        RDKafkaErrorCode::RequestTimedOut => "request_timed_out",
-        _ => "rdkafka_other",
-    }
 }
 
 // ---------------------------------------------------------------------------
@@ -89,131 +64,6 @@ pub trait SinkResult: Send + Sync {
     /// Time between batch enqueue and this event's ack completion.
     /// None if the event never entered the ack path (immediate error).
     fn elapsed(&self) -> Option<chrono::Duration>;
-}
-
-// ---------------------------------------------------------------------------
-// KafkaSinkError
-// ---------------------------------------------------------------------------
-
-/// Full-fidelity error enum capturing every failure mode in the Kafka sink.
-/// `SinkResult` trait methods derive their output from this.
-#[derive(Debug)]
-pub enum KafkaSinkError {
-    SinkNotConfigured,
-    SinkUnavailable,
-    SerializationFailed(String),
-    Produce(ProduceError),
-    Timeout,
-    TaskPanicked,
-}
-
-impl KafkaSinkError {
-    pub fn outcome(&self) -> Outcome {
-        match self {
-            Self::SinkNotConfigured => Outcome::FatalError,
-            Self::SinkUnavailable => Outcome::RetriableError,
-            Self::SerializationFailed(_) => Outcome::FatalError,
-            Self::Produce(e) => {
-                if e.is_retriable() {
-                    Outcome::RetriableError
-                } else {
-                    Outcome::FatalError
-                }
-            }
-            Self::Timeout => Outcome::Timeout,
-            Self::TaskPanicked => Outcome::RetriableError,
-        }
-    }
-
-    pub fn as_tag(&self) -> &'static str {
-        match self {
-            Self::SinkNotConfigured => "sink_not_configured",
-            Self::SinkUnavailable => "sink_unavailable",
-            Self::SerializationFailed(_) => "serialization_failed",
-            Self::Produce(e) => e.as_tag(),
-            Self::Timeout => "timeout",
-            Self::TaskPanicked => "task_panicked",
-        }
-    }
-
-    pub fn detail(&self) -> Cow<'_, str> {
-        match self {
-            Self::SinkNotConfigured => Cow::Borrowed("sink not configured"),
-            Self::SinkUnavailable => Cow::Borrowed("sink unavailable"),
-            Self::SerializationFailed(m) => Cow::Owned(format!("serialization failed: {m}")),
-            Self::Produce(e) => Cow::Owned(format!("{e}")),
-            Self::Timeout => Cow::Borrowed("produce timeout"),
-            Self::TaskPanicked => Cow::Borrowed("task panicked during delivery"),
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// KafkaResult
-// ---------------------------------------------------------------------------
-
-/// Kafka-specific implementation of [`SinkResult`]. Outcome is derived from
-/// the error -- no explicit outcome field.
-pub struct KafkaResult {
-    uuid_key: String,
-    error: Option<KafkaSinkError>,
-    enqueued_at: DateTime<Utc>,
-    completed_at: Option<DateTime<Utc>>,
-}
-
-impl KafkaResult {
-    pub(crate) fn ok(uuid_key: String, enqueued_at: DateTime<Utc>) -> Self {
-        Self {
-            uuid_key,
-            error: None,
-            enqueued_at,
-            completed_at: None,
-        }
-    }
-
-    pub(crate) fn err(uuid_key: String, error: KafkaSinkError, enqueued_at: DateTime<Utc>) -> Self {
-        Self {
-            uuid_key,
-            error: Some(error),
-            enqueued_at,
-            completed_at: None,
-        }
-    }
-
-    pub(crate) fn with_completed_at(mut self, t: DateTime<Utc>) -> Self {
-        self.completed_at = Some(t);
-        self
-    }
-
-    pub fn error(&self) -> Option<&KafkaSinkError> {
-        self.error.as_ref()
-    }
-}
-
-impl SinkResult for KafkaResult {
-    fn key(&self) -> &str {
-        &self.uuid_key
-    }
-
-    fn outcome(&self) -> Outcome {
-        match &self.error {
-            None => Outcome::Success,
-            Some(e) => e.outcome(),
-        }
-    }
-
-    fn cause(&self) -> Option<&'static str> {
-        self.error.as_ref().map(|e| e.as_tag())
-    }
-
-    fn detail(&self) -> Option<Cow<'_, str>> {
-        self.error.as_ref().map(|e| e.detail())
-    }
-
-    fn elapsed(&self) -> Option<chrono::Duration> {
-        self.completed_at
-            .map(|t| t.signed_duration_since(self.enqueued_at))
-    }
 }
 
 // ---------------------------------------------------------------------------
